% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fsolar}
\alias{fsolar}
\title{Failure Function (probability of failure)}
\usage{
fsolar(time, c, m)
}
\arguments{
\item{time}{[numeric] number of years passed}

\item{c}{[numeric] scale factor of Weibull distribution (1/lambda)}

\item{m}{[numeric] shape factor of Weibull distribution}
}
\description{
Failure Function (probability of failure)
}
