% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lifetime_dist}
\alias{lifetime_dist}
\title{Graphing Lifetime Distribution}
\usage{
lifetime_dist(lambda, num_of_panels, age, m)
}
\arguments{
\item{lambda}{[numeric] mean failure rate}

\item{age}{[numeric] age of solar panels in years}

\item{m}{[numeric] shape factor of Weibull distribution}

\item{num_of_pannels}{[numeric] number of solar panels on the farm}
}
\description{
Graphing Lifetime Distribution
}
\note{
Dependency: `fsolar()` function and `tsolar()` function
}
