% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mttf}
\alias{mttf}
\title{Mean Time to Failure}
\usage{
mttf(lambda, m)
}
\arguments{
\item{lambda}{[numeric] mean failure rate}

\item{m}{[numeric] shape factor of Weibull distribution}
}
\description{
Mean Time to Failure
}
