% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{solar_flux_from_csv}
\alias{solar_flux_from_csv}
\title{Calculating Solar Flux from csv}
\usage{
solar_flux_from_csv(filename, L, beta, gamma)
}
\arguments{
\item{filename}{[String] intended name of file}

\item{L}{[numeric] latitude of location}

\item{beta}{[numeric] tilt angle of the solar panels in degrees}

\item{gamma}{[numeric] azimuth angle of the solar panels in degrees}
}
\description{
Calculating Solar Flux from csv
}
