% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{solar_spc}
\alias{solar_spc}
\title{Solar Statistical Process Control}
\usage{
solar_spc(data)
}
\arguments{
\item{data}{[data frame] file of solar flux over time for the solar farm}
}
\description{
Solar Statistical Process Control
}
