% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tsolar}
\alias{tsolar}
\title{Time until Failures}
\usage{
tsolar(failure_prob, c, age, m)
}
\arguments{
\item{failure_prob}{[numeric] failure probability}

\item{c}{[numeric] scale factor of Weibull distribution (1/lambda)}

\item{age}{[numeric] age of solar panels in years}

\item{m}{[numeric] shape factor of Weibull distribution}
}
\description{
Time until Failures
}
